pyproject.toml
[tool.poetry]
name = "fasttower"
version = "0.1.6"
description = "A Django-like framework built on FastAPI"
authors = ["Alexander Ibragmov <sasha.2000ibr@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/pysashapy/fasttower"
repository = "https://github.com/pysashapy/fasttower.git"
keywords = [
    'python', 'api', 'framework', 'web', 'rest', 'async',
    'swagger', 'python3', 'asyncio', 'easy-web', 'starlette',
    'fastapi', 'fasttower'
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastTower",
]

packages = [
    { include = "fasttower" }
]
include = ["README.md", "LICENSE", "CHANGELOG.md"]

[tool.poetry.urls]
"Homepage" = "https://github.com/pysashapy/fasttower"
"Documentation" = "https://pysashapy.github.io/fasttower"
"Bug Tracker" = "https://github.com/pysashapy/fasttower/issues"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.5"
fastapi = "^0.115.0"
scrypt = "^0.8.24"
Jinja2 = "^3.1.4"
uvicorn = "^0.30.6"
ipython = "^8.27.0"
fastapi-debug-toolbar = "^0.6.3"
itsdangerous = "^2.2.0"

tortoise-orm = { version = "^0.21.7", optional = true }
#aerich = { git = "https://github.com/pysashapy/taerich.git", tag = "0.0.1" }
fastapi-admin = { version = "^1.0.4", optional = true }
fastapi-cache2 = "^0.2.2"
redis = "^5.2.0"
aiosmtplib = "^3.0.2"

[tool.poetry.extras]
tortoise = ["tortoise-orm", "aerich", "fastapi-admin"]

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.44"
mdx-include = "^1.4.2"
mkdocs-redirects = "^1.2.2"
markdown-include-variants = "^0.0.3"
mkdocs-macros-plugin = "^1.3.7"
mkdocstrings = { extras = ["python"], version = "^0.27.0" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tower = "fasttower.cli:app"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.mypy]
pretty = true
ignore_missing_imports = true

[tool.ruff.lint]
ignore = ['E501']
